Update the scroll view content size
This also triggers `setNeedsLayout` and `layoutIfNeeded` whenever a
change is made in the scrollview's view hierarchy. In the previous
implementation `setNeedsLayout` was called implicitly because we were
modifying the view hierarchy of the view. The view now adds subviews to
the scrollview, so it's never marked as needing layout anymore.